+++ composer.json
@@ -1,15 +1,20 @@
 {
-    "name": "moorscode/wp-multi-object-cache",
-    "description": "Object Cache allowing for multiple cache implementations",
-    "type": "wordpress-muplugin",
-    "require": {
-        "cache/cache": "^0.4.0"
-    },
-    "license": "GPLv3",
-    "authors": [
-        {
-            "name": "Jip Moors",
-            "email": "jhp.moors@gmail.com"
-        }
-    ]
+  "name": "moorscode/wp-multi-object-cache",
+  "description": "Object Cache allowing for multiple cache implementations",
+  "type": "wordpress-muplugin",
+  "require": {
+    "cache/cache": "^0.4.0"
+  },
+  "license": "GPLv3",
+  "authors": [
+    {
+      "name": "Jip Moors",
+      "email": "jhp.moors@gmail.com"
+    }
+  ],
+  "autoload": {
+    "psr-4": {
+      "MultiObjectCache\\Cache\\": "src/"
+    }
+  }
 }
