+++ src/PoolGroupConnector.php
